+++ examples/root_ca.rs
@@ -1,4 +1,4 @@
-use centerdevice::{CenterDevice, ClientBuilder, ClientCredentials, Token};
+use centerdevice::{CenterDevice, Certificate, ClientBuilder, ClientCredentials, Token};
 
 use std::env;
 
@@ -19,18 +19,26 @@ fn main() {
         .expect("Environment variable 'CENTERDEVICE_REFRESH_TOKEN' is not set.")
         .to_string_lossy()
         .to_string();
-    let document_ids = env::var_os("CENTERDEVICE_DOCUMENT_IDS")
-        .expect("Environment variable 'CENTERDEVICE_DOCUMENT_IDS' is not set.")
+    let base_url = env::var_os("CENTERDEVICE_BASE_URL")
+        .expect("Environment variable 'CENTERDEVICE_BASE_URL' is not set.")
+        .to_string_lossy()
+        .to_string();
+    let root_ca_file = env::var_os("CENTERDEVICE_ROOT_CA_FILE")
+        .expect("Environment variable 'CENTERDEVICE_ROOT_CA_FILE' is not set.")
         .to_string_lossy()
         .to_string();
 
-    let ids: Vec<&str> = document_ids.split(',').collect();
+    let pem = std::fs::read(root_ca_file).expect("Failed to read Root CA pem file");
+    let certificate = Certificate::from_pem(&pem).expect("Failed to parse Root CA");
 
     let client_credentials = ClientCredentials::new(&client_id, &client_secret);
     let token = Token::new(access_token, refresh_token);
-    let client = ClientBuilder::new("centerdevice.de", client_credentials).build_with_token(token);
 
-    client.delete_documents(&ids).expect("Download failed");
+    let client = ClientBuilder::new(&base_url, client_credentials)
+        .add_root_certificate(certificate)
+        .build_with_token(token);
 
-    println!("Ok");
+    let token = client.refresh_access_token().expect("Search failed.");
+
+    println!("Refreshed Access Token: '{:#?}'", token);
 }