+++ src/lib.rs
@@ -3,6 +3,7 @@ pub mod errors;
 pub mod utils;
 
 pub use crate::client::auth::Token;
+pub use reqwest::{blocking::Client as HttpClient, Certificate};
 
 use crate::{
     client::{
@@ -18,6 +19,8 @@ use crate::{
     errors::Result,
 };
 
+static APP_USER_AGENT: &str = concat!(env!("CARGO_PKG_NAME"), "/", env!("CARGO_PKG_VERSION"),);
+
 pub trait CenterDevice {
     fn refresh_access_token(&self) -> Result<Token>;
     fn search_documents(&self, search: Search) -> Result<SearchResult>;
@@ -29,15 +32,60 @@ pub trait CenterDevice {
     fn search_collections(&self, collections_query: CollectionsQuery) -> Result<CollectionsResult>;
 }
 
+pub struct ClientBuilder<'a> {
+    base_url:           &'a str,
+    client_credentials: ClientCredentials<'a>,
+    root_cert:          Option<Certificate>,
+}
+
+impl<'a> ClientBuilder<'a> {
+    pub fn new(base_url: &'a str, client_credentials: ClientCredentials<'a>) -> ClientBuilder<'a> {
+        ClientBuilder {
+            base_url,
+            client_credentials,
+            root_cert: None,
+        }
+    }
+
+    pub fn add_root_certificate(self, certificate: Certificate) -> Self {
+        Self {
+            root_cert: Some(certificate),
+            ..self
+        }
+    }
+
+    pub fn build(self) -> UnauthorizedClient<'a> {
+        let http_client = Self::build_http_client(self.root_cert);
+        Client::new(self.base_url, self.client_credentials, http_client)
+    }
+
+    pub fn build_with_token(self, token: Token) -> AuthorizedClient<'a> {
+        let http_client = Self::build_http_client(self.root_cert);
+        Client::with_token(self.base_url, self.client_credentials, token, http_client)
+    }
+
+    fn build_http_client(root_cert: Option<Certificate>) -> HttpClient {
+        let mut client_builder = reqwest::blocking::Client::builder().user_agent(APP_USER_AGENT);
+        if let Some(cert) = root_cert {
+            client_builder = client_builder.add_root_certificate(cert)
+        };
+        client_builder.build().expect("Failed to build HTTP client")
+    }
+}
+
 pub struct Client {}
 
 impl Client {
     #[allow(clippy::new_ret_no_self)]
-    pub fn new<'a>(base_url: &'a str, client_credentials: ClientCredentials<'a>) -> UnauthorizedClient<'a> {
+    pub fn new<'a>(
+        base_url: &'a str,
+        client_credentials: ClientCredentials<'a>,
+        http_client: HttpClient,
+    ) -> UnauthorizedClient<'a> {
         UnauthorizedClient {
             base_url,
             client_credentials,
-            http_client: reqwest::blocking::Client::new(),
+            http_client,
         }
     }
 
@@ -45,12 +93,13 @@ impl Client {
         base_url: &'a str,
         client_credentials: ClientCredentials<'a>,
         token: Token,
+        http_client: HttpClient,
     ) -> AuthorizedClient<'a> {
         AuthorizedClient {
             base_url,
             client_credentials,
             token,
-            http_client: reqwest::blocking::Client::new(),
+            http_client,
         }
     }
 }